separated Project class | separated scaffold_framework
