+++ classes.py
@@ -1,0 +1,6 @@
+class Project:
+    name = ""
+
+    def __init__(self, name):
+        self.name = name
+        print(f"Project name: {self.name}")
